requirements.txt
tensorflow==2.3.0
pandas==1.1.0
numpy==1.18.5
Pillow==7.2.0
streamlit==0.67.1
